Fix: Avoid empty string as body of task message when posting tasks.
`json_encode($body)` could return empty string, so the task's JSON
could contain `"body": ""`, which is catastrophic - Celery can't decode
that message and crashes - and the message is kept in the queue, causing
crashes again and again.

This could happen, for example, if any of the task arguments was a string
that contained badly formed Unicode characters.

Now we:
- Replace (tell json_encode() to) badly formed Unicode with "\ufffd"
(REPLACEMENT CHARACTER).
- For other problems with json_encoding we're strict and we'll throw
  `JsonException`.